+++ src/content/projects/you-wouldn-t-download-a-hedgehog.mdoc
@@ -68,19 +68,3 @@ Educational outreach programs play a vital role in spreading awareness about pro
 Here's a caption for this image of a downlaoded hedgehog
 {% /Column %}
 {% /MultiColumn %}
-
-{% MultiColumn
-   justifyContent="center"
-   alignItems="center"
-   flexDirection="row"
-   gap="20px" %}
-{% Column targetWidth="300px" flexGrow=false %}
-### This is a Downloaded Hedgehog
-
-![](../../assets/images/pages/you-wouldn-t-download-a-hedgehog/dreamstime_s_280673852.jpg)
-{% /Column %}
-
-{% Column targetWidth="300px" flexGrow=false customClass="custom-class" %}
-Here's a caption for this image of a downlaoded hedgehog
-{% /Column %}
-{% /MultiColumn %}
